refactor: move LendingEngine interface definition to state_processor_viction.go, optimize
